--- data/data_80F5668.s
@@ -1,122 +1,0 @@
-        .section .rodata
-
-	.global gDungeonMusic
-gDungeonMusic: @ 80F5668
-        .2byte 0x72, 0x7d, 0x0e, 0x78, 0x0f, 0x71, 0x6f, 0x14, 0x66, 0x19, 0x68, 0x73, 0x6c, 0x16, 0x7d, 0x7d, 0x7d, 0x7d, 0x0d, 0x15, 0x0f, 0x7e, 0x21, 0x27, 0x7b, 0x7c, 0x17, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x1a, 0x6e, 0x6e, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x1a, 0x7f, 0x23, 0x23, 0x23, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d, 0x7d
-
-	.global gUnknown_80F5700
-gUnknown_80F5700: @ 80F5700
-	.incbin "baserom.gba", 0xF5700, 0xCA
-
-	.global gUnknown_80F57CA
-gUnknown_80F57CA: @ 80F57CA
-@ replacing .incbin "baserom.gba", 0x000f57ca, 0x2
-        .byte 0x0a, 0x00
-
-	.global gUnknown_80F57CC
-gUnknown_80F57CC: @ 80F57CC
-@ replacing .incbin "baserom.gba", 0x000f57cc, 0x2
-        .byte 0x14, 0x00
-
-	.global gUnknown_80F57CE
-gUnknown_80F57CE: @ 80F57CE
-@ replacing .incbin "baserom.gba", 0x000f57ce, 0x2
-        .byte 0x14, 0x00
-
-	.global gUnknown_80F57D0
-gUnknown_80F57D0: @ 80F57D0
-@ replacing .incbin "baserom.gba", 0x000f57d0, 0x2
-        .byte 0x14, 0x00
-
-	.global gUnknown_80F57D2
-gUnknown_80F57D2: @ 80F57D2
-@ replacing .incbin "baserom.gba", 0x000f57d2, 0x2
-        .byte 0x0a, 0x00
-
-	.global gUnknown_80F57D4
-gUnknown_80F57D4: @ 80F57D4
-	.incbin "baserom.gba", 0xF57D4, 0x120
-
-	.global gUnknown_80F58F4
-gUnknown_80F58F4: @ 80F58F4
-	.incbin "baserom.gba", 0xF58F4, 0x84
-
-	.global gSecretPowerTable
-gSecretPowerTable: @ 80F5978
-        .byte 0x03
-        .byte 0x00
-        .byte 0x06
-        .byte 0x07
-        .byte 0x07
-        .byte 0x06
-        .byte 0x06
-        .byte 0x07
-        .byte 0x08
-        .byte 0x06
-        .byte 0x00
-        .byte 0x05
-        .byte 0x08
-        .byte 0x08
-        .byte 0x00
-        .byte 0x08
-        .byte 0x08
-        .byte 0x07
-        .byte 0x07
-        .byte 0x07
-        .byte 0x00
-        .byte 0x05
-        .byte 0x06
-        .byte 0x08
-        .byte 0x06
-        .byte 0x00
-        .byte 0x00
-        .byte 0x08
-        .byte 0x06
-        .byte 0x01
-        .byte 0x07
-        .byte 0x07
-        .byte 0x01
-        .byte 0x01
-        .byte 0x06
-        .byte 0x08
-        .byte 0x08
-        .byte 0x07
-        .byte 0x06
-        .byte 0x06
-        .byte 0x06
-        .byte 0x00
-        .byte 0x06
-        .byte 0x06
-        .byte 0x05
-        .byte 0x08
-        .byte 0x06
-        .byte 0x06
-        .byte 0x06
-        .byte 0x03
-        .byte 0x00
-        .byte 0x00
-        .byte 0x05
-        .byte 0x02
-        .byte 0x07
-        .byte 0x06
-        .byte 0x00
-        .byte 0x08
-        .byte 0x08
-        .byte 0x06
-        .byte 0x00
-        .byte 0x07
-        .byte 0x00
-        .byte 0x00
-        .byte 0x06
-        .byte 0x01
-        .byte 0x06
-        .byte 0x06
-        .byte 0x08
-        .byte 0x08
-        .byte 0x06
-        .byte 0x08
-        .byte 0x08
-        .byte 0x03
-        .byte 0x03
-        .byte 0x06
-
