+++ src/code_806FDF4.c
@@ -23,11 +23,10 @@
 #include "text_util.h"
 #include "dungeon_util_1.h"
 #include "math.h"
-#include "dungeon_battle_data.h"
+#include "dungeon_config.h"
 #include "string_format.h"
 
 extern u8 gUnknown_202EE70[MAX_TEAM_BODY_SIZE];
-extern s16 gUnknown_80F5700[];
 extern s32 gUnknown_8106F7C[];
 
 extern u8 *gUnknown_80FA0F0[];
@@ -246,7 +245,7 @@ bool8 sub_806F660(Entity *pokemon, Entity *target)
     if (HasHeldItem(pokemon, ITEM_FRIEND_BOW)) {
         recruitRate += gFriendBowRecruitRateUpValue;
     }
-    recruitRate += gUnknown_80F5700[pokemonInfo->level];
+    recruitRate += gRecruitRateByLevel[pokemonInfo->level];
     if (rand >= recruitRate)
         return FALSE;
