+++ src/code_8077274_1.c
@@ -27,7 +27,7 @@
 #include "structs/str_dungeon.h"
 #include "dungeon_map_access.h"
 #include "structs/str_806B7F8.h"
-#include "dungeon_battle_data.h"
+#include "dungeon_config.h"
 
 extern u8 *gUnknown_80FA8BC[];
 extern u8 *gUnknown_80FA824[];
