+++ src/trap.c
@@ -24,7 +24,7 @@
 #include "structs/map.h"
 #include "structs/str_dungeon.h"
 #include "structs/str_806B7F8.h"
-#include "dungeon_battle_data.h"
+#include "dungeon_config.h"
 
 extern u8 *gTrapNames[];
 extern u8 *gUnknown_80FC5FC[];
